-- Enabling support for compiling on 32-bit architectures.     The patches were submitted by James Devine on kosmosfs-users     mailing list and are being included in the source.

+++ src/cc/chunk/ChunkServer_main.cc
@@ -58,7 +58,7 @@ string gLogDir;
 vector<string> gChunkDirs;
 
 ServerLocation gMetaServerLoc;
-size_t gTotalSpace;			// max. storage space to use
+int64_t gTotalSpace;			// max. storage space to use
 int gChunkServerClientPort;	// Port at which kfs clients connect to us
 
 Properties gProp;
