+++ src/cc/tests/KfsSeekWrite_main.cc
@@ -318,7 +318,7 @@ doRead(int fd, off_t offset, int numBytes, const char *src)
         toCompare =  min(KFS_DATA_BUF_SIZE - (start % KFS_DATA_BUF_SIZE),
                          (long long) (numBytes - bytesCompared));
 #else
-        toCompare =  min(KFS_DATA_BUF_SIZE - (start % KFS_DATA_BUF_SIZE),
+        toCompare =  min((long) (KFS_DATA_BUF_SIZE - (start % KFS_DATA_BUF_SIZE)),
                          (long) (numBytes - bytesCompared));
 #endif
         if (!compareData(kfsBuf1 + bytesCompared, src + (start % KFS_DATA_BUF_SIZE),
